Move asyncFocus from util to device.
