+++ src/components/NewThought.tsx
@@ -8,9 +8,10 @@ import { connect } from 'react-redux'
 import classNames from 'classnames'
 import { store } from '../store'
 import { MAX_DISTANCE_FROM_CURSOR } from '../constants'
-import { appendToPath, asyncFocus, createId, getTextContentFromHTML, pathToContext, unroot } from '../util'
+import { appendToPath, createId, getTextContentFromHTML, pathToContext, unroot } from '../util'
 import { getNextRank, getChildrenRanked } from '../selectors'
 import { cursorBack, createThought, setCursor } from '../action-creators'
+import asyncFocus from '../device/asyncFocus'
 import { Path, SimplePath, State } from '../@types'
 
 interface NewThoughtProps {
