+++ src/shortcuts/delete.native.tsx
@@ -2,11 +2,12 @@ import React from 'react'
 // import { Key } from 'ts-key-enum'
 // import { isTouch } from '../browser'
 // import { hasChild } from '../selectors'
-// import { asyncFocus, ellipsize, headValue, isDocumentEditable, isEM, isRoot, pathToContext, setSelection } from '../util'
+// import { ellipsize, headValue, isDocumentEditable, isEM, isRoot, pathToContext, setSelection } from '../util'
 // import { alert, archiveThought, deleteAttribute, error } from '../action-creators'
 import { Icon as IconType, Shortcut } from '../@types'
 import Svg, { G, Path } from 'react-native-svg'
 import { Alert } from 'react-native'
+// import asyncFocus from '../device/asyncFocus'
 
 // let undoArchiveTimer: number // eslint-disable-line fp/no-let
