+++ src/shortcuts/deleteEmptyThoughtOrOutdent.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Key } from 'ts-key-enum'
-import { asyncFocus, ellipsize, headValue, isDivider, isDocumentEditable, parentOf, pathToContext } from '../util'
+import { ellipsize, headValue, isDivider, isDocumentEditable, parentOf, pathToContext } from '../util'
 import {
   getChildren,
   getThoughtBefore,
@@ -15,6 +15,7 @@ import {
 import { HOME_PATH } from '../constants'
 import { isTouch } from '../browser'
 import { alert, deleteEmptyThought as deleteEmptyThoughtActionCreator, error, outdent } from '../action-creators'
+import asyncFocus from '../device/asyncFocus'
 import { Icon as IconType, Shortcut, State, Thunk } from '../@types'
 
 /** Returns true if the cursor is on an empty though or divider that can be deleted. */
