+++ src/util/index.ts
@@ -4,7 +4,6 @@ export { addContext } from './addContext'
 export { addEmojiSpace } from './addEmojiSpace'
 export { ancestors } from './ancestors'
 export { appendToPath } from './appendToPath'
-export { asyncFocus } from './asyncFocus'
 export { checkIfPathShareSubcontext } from './checkIfPathShareSubcontext'
 export { clearSelection } from './clearSelection'
 export { compareByRank } from './compareByRank'
